requirements.txt
certifi==2022.9.14
charset-normalizer==2.1.1
click==8.1.3
dlib==19.24.0
face-recognition==1.3.0
face-recognition-models==0.3.0
gTTS==2.2.3
idna==3.4
importlib-metadata==4.12.0
imutils==0.5.4
joblib==1.2.0
numpy==1.21.0
opencv-python==4.5.3.56
opencv-python-headless==4.5.3.56
Pillow==9.2.0
pkg-resources==0.0.0
PyAudio==0.2.12
pydub==0.25.1
pyttsx3==2.90
requests==2.28.1
RPi.GPIO==0.7.1
scikit-learn==0.24.2
scipy==1.7.3
six==1.16.0
SpeechRecognition==3.8.1
threadpoolctl==3.1.0
typing-extensions==4.3.0
urllib3==1.26.12
zipp==3.8.1
